+++ packages/core/src/linker/component_factory_resolver.ts
@@ -6,12 +6,10 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {Injector} from '../di/injector';
 import {Type} from '../interface/type';
 import {stringify} from '../util/stringify';
 
-import {ComponentFactory, ComponentRef} from './component_factory';
-import {NgModuleRef} from './ng_module_factory';
+import {ComponentFactory} from './component_factory';
 
 export function noComponentFactoryError(component: Function) {
   const error = Error(`No component factory found for ${
