+++ packages/core/src/render3/component.ts
@@ -6,236 +6,184 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-// We are temporarily importing the existing viewEngine from core so we can be sure we are
-// correctly implementing its interfaces for backwards compatibility.
 import {Injector} from '../di/injector';
-import {Sanitizer} from '../sanitization/sanitizer';
-import {assertDefined, assertIndexInRange} from '../util/assert';
+import {EnvironmentInjector, getNullInjector} from '../di/r3_injector';
+import {Type} from '../interface/type';
+import {ComponentRef} from '../linker/component_factory';
 
-import {diPublicInInjector, getOrCreateNodeInjectorForNode} from './di';
-import {throwProviderNotFoundError} from './errors_di';
-import {registerPostOrderHooks} from './hooks';
-import {addToViewTree, CLEAN_PROMISE, createLView, getOrCreateTComponentView, getOrCreateTNode, initTNodeFlags, instantiateRootComponent, invokeHostBindingsInCreationMode, markAsComponentHost, registerHostBindingOpCodes} from './instructions/shared';
-import {ComponentDef, RenderFlags} from './interfaces/definition';
-import {TElementNode, TNodeType} from './interfaces/node';
-import {PlayerHandler} from './interfaces/player';
-import {Renderer, RendererFactory} from './interfaces/renderer';
-import {RElement} from './interfaces/renderer_dom';
-import {CONTEXT, HEADER_OFFSET, LView, LViewFlags, RootContext, RootContextFlags, TVIEW} from './interfaces/view';
-import {writeDirectClass, writeDirectStyle} from './node_manipulation';
-import {getCurrentTNode, getLView, setSelectedIndex} from './state';
-import {computeStaticStyling} from './styling/static_styling';
-import {setUpAttributes} from './util/attrs_utils';
-import {defaultScheduler} from './util/misc_utils';
-import {getRootContext} from './util/view_traversal_utils';
-
-
-
-/** Options that control how the component should be bootstrapped. */
-export interface CreateComponentOptions {
-  /** Which renderer factory to use. */
-  rendererFactory?: RendererFactory;
-
-  /** A custom sanitizer instance */
-  sanitizer?: Sanitizer;
-
-  /** A custom animation player handler */
-  playerHandler?: PlayerHandler;
-
-  /**
-   * Host element on which the component will be bootstrapped. If not specified,
-   * the component definition's `tag` is used to query the existing DOM for the
-   * element to bootstrap.
-   */
-  host?: RElement|string;
-
-  /** Module injector for the component. If unspecified, the injector will be NULL_INJECTOR. */
-  injector?: Injector;
-
-  /**
-   * List of features to be applied to the created component. Features are simply
-   * functions that decorate a component with a certain behavior.
-   *
-   * Typically, the features in this list are features that cannot be added to the
-   * other features list in the component definition because they rely on other factors.
-   *
-   * Example: `LifecycleHooksFeature` is a function that adds lifecycle hook capabilities
-   * to root components in a tree-shakable way. It cannot be added to the component
-   * features list because there's no way of knowing when the component will be used as
-   * a root component.
-   */
-  hostFeatures?: HostFeature[];
-
-  /**
-   * A function which is used to schedule change detection work in the future.
-   *
-   * When marking components as dirty, it is necessary to schedule the work of
-   * change detection in the future. This is done to coalesce multiple
-   * {@link markDirty} calls into a single changed detection processing.
-   *
-   * The default value of the scheduler is the `requestAnimationFrame` function.
-   *
-   * It is also useful to override this function for testing purposes.
-   */
-  scheduler?: (work: () => void) => void;
-}
-
-/** See CreateComponentOptions.hostFeatures */
-type HostFeature = (<T>(component: T, componentDef: ComponentDef<T>) => void);
-
-// TODO: A hack to not pull in the NullInjector from @angular/core.
-export const NULL_INJECTOR: Injector = {
-  get: (token: any, notFoundValue?: any) => {
-    throwProviderNotFoundError(token, 'NullInjector');
-  }
-};
+import {ComponentFactory} from './component_ref';
+import {getComponentDef} from './definition';
+import {assertComponentDef} from './errors';
 
 /**
- * Creates the root component view and the root component node.
+ * Creates a `ComponentRef` instance based on provided component type and a set of options.
  *
- * @param rNode Render host element.
- * @param def ComponentDef
- * @param rootView The parent view where the host node is stored
- * @param rendererFactory Factory to be used for creating child renderers.
- * @param hostRenderer The current renderer
- * @param sanitizer The sanitizer, if provided
+ * @usageNotes
  *
- * @returns Component view created
+ * The example below demonstrates how the `createComponent` function can be used
+ * to create an instance of a ComponentRef dynamically and attach it to an ApplicationRef,
+ * so that it gets included into change detection cycles.
+ *
+ * Note: the example uses standalone components, but the function can also be used for
+ * non-standalone components (declared in an NgModule) as well.
+ *
+ * ```typescript
+ * @Component({
+ *   standalone: true,
+ *   template: `Hello {{ name }}!`
+ * })
+ * class HelloComponent {
+ *   name = 'Angular';
+ * }
+ *
+ * @Component({
+ *   standalone: true,
+ *   template: `<div id="hello-component-host"></div>`
+ * })
+ * class RootComponent {}
+ *
+ * // Bootstrap an application.
+ * const applicationRef = await bootstrapApplication(RootComponent);
+ *
+ * // Locate a DOM node that would be used as a host.
+ * const host = document.getElementById('hello-component-host');
+ *
+ * // Get an `EnvironmentInjector` instance from the `ApplicationRef`.
+ * const environmentInjector = applicationRef.injector;
+ *
+ * // We can now create a `ComponentRef` instance.
+ * const componentRef = createComponent(HelloComponent, {host, environmentInjector});
+ *
+ * // Last step is to register the newly created ref using the `ApplicationRef` instance
+ * // to include the component view into change detection cycles.
+ * applicationRef.attachView(componentRef.hostView);
+ * ```
+ *
+ * @param component Component class reference.
+ * @param options Set of options to use:
+ *  * `environmentInjector`: An `EnvironmentInjector` instance to be used for the component, see
+ * additional info about it at https://angular.io/guide/standalone-components#environment-injectors.
+ *  * `hostElement` (optional): A DOM node that should act as a host node for the component. If not
+ * provided, Angular creates one based on the tag name used in the component selector (and falls
+ * back to using `div` if selector doesn't have tag name info).
+ *  * `elementInjector` (optional): An `ElementInjector` instance, see additional info about it at
+ * https://angular.io/guide/hierarchical-dependency-injection#elementinjector.
+ *  * `projectableNodes` (optional): A list of DOM nodes that should be projected through
+ *                      [`<ng-content>`](api/core/ng-content) of the new component instance.
+ * @returns ComponentRef instance that represents a given Component.
+ *
+ * @publicApi
  */
-export function createRootComponentView(
-    rNode: RElement|null, def: ComponentDef<any>, rootView: LView, rendererFactory: RendererFactory,
-    hostRenderer: Renderer, sanitizer?: Sanitizer|null): LView {
-  const tView = rootView[TVIEW];
-  const index = HEADER_OFFSET;
-  ngDevMode && assertIndexInRange(rootView, index);
-  rootView[index] = rNode;
-  // '#host' is added here as we don't know the real host DOM name (we don't want to read it) and at
-  // the same time we want to communicate the debug `TNode` that this is a special `TNode`
-  // representing a host element.
-  const tNode: TElementNode = getOrCreateTNode(tView, index, TNodeType.Element, '#host', null);
-  const mergedAttrs = tNode.mergedAttrs = def.hostAttrs;
-  if (mergedAttrs !== null) {
-    computeStaticStyling(tNode, mergedAttrs, true);
-    if (rNode !== null) {
-      setUpAttributes(hostRenderer, rNode, mergedAttrs);
-      if (tNode.classes !== null) {
-        writeDirectClass(hostRenderer, rNode, tNode.classes);
-      }
-      if (tNode.styles !== null) {
-        writeDirectStyle(hostRenderer, rNode, tNode.styles);
-      }
-    }
-  }
-
-  const viewRenderer = rendererFactory.createRenderer(rNode, def);
-  const componentView = createLView(
-      rootView, getOrCreateTComponentView(def), null,
-      def.onPush ? LViewFlags.Dirty : LViewFlags.CheckAlways, rootView[index], tNode,
-      rendererFactory, viewRenderer, sanitizer || null, null, null);
-
-  if (tView.firstCreatePass) {
-    diPublicInInjector(getOrCreateNodeInjectorForNode(tNode, rootView), tView, def.type);
-    markAsComponentHost(tView, tNode);
-    initTNodeFlags(tNode, rootView.length, 1);
-  }
-
-  addToViewTree(rootView, componentView);
-
-  // Store component view at node index, with node as the HOST
-  return rootView[index] = componentView;
+export function createComponent<C>(component: Type<C>, options: {
+  environmentInjector: EnvironmentInjector,
+  hostElement?: Element,
+  elementInjector?: Injector,
+  projectableNodes?: Node[][],
+}): ComponentRef<C> {
+  ngDevMode && assertComponentDef(component);
+  const componentDef = getComponentDef(component)!;
+  const elementInjector = options.elementInjector || getNullInjector();
+  const factory = new ComponentFactory<C>(componentDef);
+  return factory.create(
+      elementInjector, options.projectableNodes, options.hostElement, options.environmentInjector);
 }
 
 /**
- * Creates a root component and sets it up with features and host bindings. Shared by
- * renderComponent() and ViewContainerRef.createComponent().
+ * An interface that describes the subset of component metadata
+ * that can be retrieved using the `reflectComponentType` function.
+ *
+ * @publicApi
  */
-export function createRootComponent<T>(
-    componentView: LView, componentDef: ComponentDef<T>, rootLView: LView, rootContext: RootContext,
-    hostFeatures: HostFeature[]|null): any {
-  const tView = rootLView[TVIEW];
-  // Create directive instance with factory() and store at next index in viewData
-  const component = instantiateRootComponent(tView, rootLView, componentDef);
-
-  rootContext.components.push(component);
-  componentView[CONTEXT] = component;
-
-  if (hostFeatures !== null) {
-    for (const feature of hostFeatures) {
-      feature(component, componentDef);
-    }
-  }
-
-  // We want to generate an empty QueryList for root content queries for backwards
-  // compatibility with ViewEngine.
-  if (componentDef.contentQueries) {
-    const tNode = getCurrentTNode()!;
-    ngDevMode && assertDefined(tNode, 'TNode expected');
-    componentDef.contentQueries(RenderFlags.Create, component, tNode.directiveStart);
-  }
-
-  const rootTNode = getCurrentTNode()!;
-  ngDevMode && assertDefined(rootTNode, 'tNode should have been already created');
-  if (tView.firstCreatePass &&
-      (componentDef.hostBindings !== null || componentDef.hostAttrs !== null)) {
-    setSelectedIndex(rootTNode.index);
-
-    const rootTView = rootLView[TVIEW];
-    registerHostBindingOpCodes(
-        rootTView, rootTNode, rootLView, rootTNode.directiveStart, rootTNode.directiveEnd,
-        componentDef);
-
-    invokeHostBindingsInCreationMode(componentDef, component);
-  }
-  return component;
+export interface ComponentMirror<C> {
+  /**
+   * The component's HTML selector.
+   */
+  get selector(): string;
+  /**
+   * The type of component the factory will create.
+   */
+  get type(): Type<C>;
+  /**
+   * The inputs of the component.
+   */
+  get inputs(): ReadonlyArray<{propName: string, templateName: string}>;
+  /**
+   * The outputs of the component.
+   */
+  get outputs(): ReadonlyArray<{propName: string, templateName: string}>;
+  /**
+   * Selector for all <ng-content> elements in the component.
+   */
+  get ngContentSelectors(): ReadonlyArray<string>;
+  /**
+   * Whether this component is marked as standalone.
+   * Note: an extra flag, not present in `ComponentFactory`.
+   */
+  get isStandalone(): boolean;
 }
 
+/**
+ * Creates an object that allows to retrieve component metadata.
+ *
+ * @usageNotes
+ *
+ * The example below demonstrates how to use the function and how the fields
+ * of the returned object map to the component metadata.
+ *
+ * ```typescript
+ * @Component({
+ *   standalone: true,
+ *   selector: 'foo-component',
+ *   template: `
+ *     <ng-content></ng-content>
+ *     <ng-content select="content-selector-a"></ng-content>
+ *   `,
+ * })
+ * class FooComponent {
+ *   @Input('inputName') inputPropName: string;
+ *   @Output('outputName') outputPropName = new EventEmitter<void>();
+ * }
+ *
+ * const mirror = reflectComponentType(FooComponent);
+ * expect(mirror.type).toBe(FooComponent);
+ * expect(mirror.selector).toBe('foo-component');
+ * expect(mirror.isStandalone).toBe(true);
+ * expect(mirror.inputs).toEqual([{propName: 'inputName', templateName: 'inputPropName'}]);
+ * expect(mirror.outputs).toEqual([{propName: 'outputName', templateName: 'outputPropName'}]);
+ * expect(mirror.ngContentSelectors).toEqual([
+ *   '*',                 // first `<ng-content>` in a template, the selector defaults to `*`
+ *   'content-selector-a' // second `<ng-content>` in a template
+ * ]);
+ * ```
+ *
+ * @param component Component class reference.
+ * @returns An object that allows to retrieve component metadata.
+ *
+ * @publicApi
+ */
+export function reflectComponentType<C>(component: Type<C>): ComponentMirror<C>|null {
+  const componentDef = getComponentDef(component);
+  if (!componentDef) return null;
 
-export function createRootContext(
-    scheduler?: (workFn: () => void) => void, playerHandler?: PlayerHandler|null): RootContext {
+  const factory = new ComponentFactory<C>(componentDef);
   return {
-    components: [],
-    scheduler: scheduler || defaultScheduler,
-    clean: CLEAN_PROMISE,
-    playerHandler: playerHandler || null,
-    flags: RootContextFlags.Empty
+    get selector(): string {
+      return factory.selector;
+    },
+    get type(): Type<C> {
+      return factory.componentType;
+    },
+    get inputs(): ReadonlyArray<{propName: string, templateName: string}> {
+      return factory.inputs;
+    },
+    get outputs(): ReadonlyArray<{propName: string, templateName: string}> {
+      return factory.outputs;
+    },
+    get ngContentSelectors(): ReadonlyArray<string> {
+      return factory.ngContentSelectors;
+    },
+    get isStandalone(): boolean {
+      return componentDef.standalone;
+    },
   };
-}
-
-/**
- * Used to enable lifecycle hooks on the root component.
- *
- * Include this feature when calling `renderComponent` if the root component
- * you are rendering has lifecycle hooks defined. Otherwise, the hooks won't
- * be called properly.
- *
- * Example:
- *
- * ```
- * renderComponent(AppComponent, {hostFeatures: [LifecycleHooksFeature]});
- * ```
- */
-export function LifecycleHooksFeature(): void {
-  const tNode = getCurrentTNode()!;
-  ngDevMode && assertDefined(tNode, 'TNode is required');
-  registerPostOrderHooks(getLView()[TVIEW], tNode);
-}
-
-/**
- * Wait on component until it is rendered.
- *
- * This function returns a `Promise` which is resolved when the component's
- * change detection is executed. This is determined by finding the scheduler
- * associated with the `component`'s render tree and waiting until the scheduler
- * flushes. If nothing is scheduled, the function returns a resolved promise.
- *
- * Example:
- * ```
- * await whenRendered(myComponent);
- * ```
- *
- * @param component Component to wait upon
- * @returns Promise which resolves when the component is rendered.
- */
-export function whenRendered(component: any): Promise<null> {
-  return getRootContext(component).clean;
 }
