+++ packages/core/test/bundling/hello_world/bundle.golden_symbols.json
@@ -351,9 +351,6 @@
     "name": "createPlatformFactory"
   },
   {
-    "name": "createRootComponentView"
-  },
-  {
     "name": "createTView"
   },
   {
