+++ packages/core/test/bundling/standalone_bootstrap/bundle.golden_symbols.json
@@ -417,9 +417,6 @@
     "name": "createNodeInjector"
   },
   {
-    "name": "createRootComponentView"
-  },
-  {
     "name": "createTView"
   },
   {
